Move theme assets to generator repository
Description:
All of the necessary files from the "Goa" theme are moved to the website
generator repo so that the files can be modified as needed. Several lines of
code have been adjusted.

Also the Goa default browser header icon is replaced by the default Hugo icon
(this may be adjusted later on).

Build well